pyproject.toml
[project]
name = "hugging-face-ai-agent-course"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "ddgs>=9.12.1",
    "duckduckgo-search>=8.0.0",
    "geopandas>=1.1.3",
    "huggingface-hub>=1.8.0",
    "kaleido>=1.2.0",
    "markdownify>=0.13.1",
    "plotly>=6.6.0",
    "pytz>=2026.1.post1",
    "pyyaml>=6.0.3",
    "python-dotenv>=1.0.0",
    "requests>=2.33.0",
    "shapely>=2.1.2",
    "smolagents[gradio,litellm]>=1.22.0",
]

[project.scripts]
start = "hugging_face_ai_agent_course:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
package = true

[dependency-groups]
dev = [
    "black>=26.3.1",
    "pre-commit>=4.5.1",
]

[tool.black]
line-length = 88
target-version = ['py312']
include = '\.pyi?$'
